Gitmetrics v0.0.1 Gitmetrics.Managment

Modulo per github api

Functions

api_call(arg)

api_call(arg, client)

Chiamata principale della lista di tutti issue presenti da qualle poi verano costruite grafici nel caso la trova gia nella cash la carica semplicemente. due Implementazione sia per autorizzati che non.

can_i_send?(list)

get_data(issue)

inserisci dati delle issue in una mappa dalla chiamata api_call

init_url(url)

init_url(String.t) :: {String.t, String.t}

Prendo dal url solo ultime due parti -> organizzazione e ripo

send_close_time(list)

Restituisce lista con dentro mappate il numero del issue e suo diff del tempo di chiusura Tempo.creazione - Tempo.chiusura

send_info_status(list)

Restituisce lista di stato con tutte issue (chiuse senza commento, non commentate e senza un label)

send_issues_time(list, org, repo)

Ritorna lista del tempo dalla creazione alla prima risposta

send_issues_time(list, org, repo, client)

send_metrics_lista(list)

restituisce la lista con descrizioni titolo e body delle issue

send_metrics_stato(list)

Restituisce totale di issue e aperte, chiuse e in generale